requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autocorrect==2.6.1
backoff==2.2.1
blis==0.7.9
build==0.10.0
CacheControl==0.12.11
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
cohere==4.0.4
confection==0.0.4
crashtest==0.4.1
cymem==2.0.7
dataclasses-json==0.5.7
distlib==0.3.6
dulwich==0.21.3
enchant==0.0.1
fastapi==0.94.0
filelock==3.9.0
frozenlist==1.3.3
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
inexactsearch==1.0.2
iniconfig==2.0.0
installer==0.6.0
jaraco.classes==3.2.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keyring==23.13.1
langchain==0.0.115
langcodes==3.3.0
lockfile==0.12.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==2.0.0rc1
pathy==0.10.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
preshed==3.0.8
ptyprocess==0.7.0
pyahocorasick==1.4.2
pycparser==2.21
pydantic==1.10.6
pyenchant==3.2.2
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pyspellchecker==0.7.1
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
scikit-learn==1.2.2
scipy==1.10.1
shellingham==1.5.0.post1
silpa-common==0.3
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soundex==1.1.3
spacy==4.0.0.dev0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spellchecker==0.4
SQLAlchemy==1.4.47
srsly==2.4.6
starlette==0.26.1
tenacity==8.2.2
thinc==9.0.0.dev2
threadpoolctl==3.1.0
tomlkit==0.11.6
tqdm==4.65.0
trove-classifiers==2023.3.9
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.0
uvloop==0.17.0
virtualenv==20.21.0
wasabi==1.1.1
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
xattr==0.10.1
yarl==1.8.2
zipp==3.15.0
